pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mba-rr"
version = "0.1.0"
description = "Sample Python CLI scaffold for the mba-rr workspace."
readme = "README.md"
authors = [
    { name = "Workspace Owner" },
]
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "vaderSentiment>=3.3.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.3",
    "black>=24.2",
]

[project.scripts]
mba-rr = "mba_rr.cli:main"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "ruff>=0.3",
    "black>=24.2",
]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
target-version = "py311"
select = [
    "E",
    "F",
    "I",
    "UP",
    "B",
    "C4",
    "SIM",
    "N",
    "ANN",
    "RUF",
]
fix = true
unsafe-fixes = false
src = ["src"]

[tool.ruff.lint]
ignore = []

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-ra"
